Migrated reducers, minor refactor
