+++ app/reducers/settingsReducer.js
@@ -1,0 +1,7 @@
+const settings_data = {};
+
+const settingsReducer = (state=settings_data, action) => {
+  return state;
+}
+
+export { settingsReducer };
