+++ app/store/initialization.js
@@ -3,34 +3,19 @@ import axios from "axios";
 import logger from "redux-logger";
 import thunk from "redux-thunk";
 
-import { transaction_data } from '../components/Transactions/TransactionData';
-
-// TODO: move to separate file
-const settings_data = {};
-
-// TODO: move to separate file
-const transactionsReducer = (state=transaction_data, action) => {
-  switch (action.type) {
-    case "ADD_TRANSACTION": {
-      return [
-        ...state,
-        action.payload
-      ];
-    }
-  }
-  return state;
-}
-
-// TODO: move to separate file
-const settingsReducer = (state=settings_data, action) => {
-  return state;
-}
+import { transactionsReducer } from '../reducers/transactionsReducer';
+import { settingsReducer } from '../reducers/settingsReducer';
 
 const reducers = combineReducers({
   settings: settingsReducer,
   transactions: transactionsReducer,
 })
 
+/*
+ * React middleware works on the middle of the flow of a dispatched event and can
+ * modify, read, or even cancel it. Logger is helpful for debugging purposes, and
+ * redux-thunk allows returning actions as functions for asynchronous actions
+ */
 const middleware = applyMiddleware(thunk, logger());
 
 const store = createStore(reducers, middleware);
@@ -40,6 +25,6 @@ store.subscribe(() => {
 })
 
 // store.dispatch
-store.dispatch({type: "ADD_TRANSACTION", payload: {name: "New Transaction", date: "Date", price: -124}});
+// store.dispatch({type: "ADD_TRANSACTION", payload: {name: "New Transaction", date: "Date", price: -124}});
 
 export { store };
