+++ app/styles/globals/_variables.scss
@@ -8,4 +8,5 @@ $text-color: #666;
   list-style-type: none;
   padding-left: 0;
   margin-bottom: 0;
+  margin-top: 0;
 }
