+++ app/styles/layout/_layout.scss
@@ -1,5 +1,6 @@
 body {
   background-color: $background-color;
+  margin: 0;
 }
 .no-margin {
   margin: 0;
