+++ app/styles/layout/_nav.scss
@@ -36,4 +36,5 @@
   text-align: center;
   font-size: 22px;
   line-height: 50px;
+  z-index: 5;
 }
